Refactor app bootstrap and Supabase workflow
