+++ .github/workflows/ci.yml
@@ -14,3 +14,4 @@ jobs:
       - run: bun install --frozen-lockfile
       - run: bun run check
       - run: bun run test
+      - run: bun run test:http
