+++ .tickets/ar-g8y7.md
@@ -1,0 +1,19 @@
+---
+id: ar-g8y7
+status: open
+deps: [ar-k4qi, ar-3hsi, ar-xdcu, ar-g1z6, ar-5kph, ar-m8ai, ar-ezes, ar-ee08, ar-mpbi, ar-naq0]
+links: []
+created: 2026-07-10T21:28:56Z
+type: epic
+priority: 0
+assignee: David Torres
+tags: [architecture, contributor-experience, quality]
+---
+# Sustainability hardening program
+
+Address the maintainability, test reliability, CI coverage, service-boundary, and operational gaps identified in the July 2026 project review.
+
+## Acceptance Criteria
+
+All child tickets are closed; documented validation tiers are reproducible from a clean checkout; CI enforces the agreed checks.
+
