+++ .tickets/ar-naq0.md
@@ -1,0 +1,19 @@
+---
+id: ar-naq0
+status: open
+deps: [ar-xdcu]
+links: []
+created: 2026-07-10T21:28:57Z
+type: task
+priority: 2
+assignee: David Torres
+tags: [operations, documentation]
+---
+# Document and automate operational runbooks
+
+Deployment/runtime compatibility, migration rollout/repair, backup restore, and ownership practices remain underspecified. The backup script is hard-coded to one Supabase project.
+
+## Acceptance Criteria
+
+Repository documents deployment config, runtime support, migration and rollback/repair, backup and restore drills, and ownership/review expectations; backup tooling is environment-configurable and safely validated.
+
