+++ app.js
@@ -1,0 +1,96 @@
+const express = require('express');
+const exphbs = require('express-handlebars');
+const helpers = require('handlebars-helpers')();
+const {
+  times, date_tz, calendar_link, getTotalV1MissionsNeeded, getTotalV2MissionsNeeded,
+  setVariable, encodeURIComponentH, dump, videoEmbed, isSupportedVideoUrl, substring,
+  concat, effectiveRulesVersion, wordCount, perksForAbility, nextPerkPosition, json
+} = require('./util/handlebars');
+const { renderMarkdown } = require('./util/markdown');
+const { sendError } = require('./util/http-error');
+const range = require('handlebars-helper-range');
+const path = require('path');
+
+const homeRoutes = require('./routes/home');
+const authRoutes = require('./routes/auth');
+const charactersRoutes = require('./routes/characters');
+const lfgRoutes = require('./routes/lfg');
+const profileRoutes = require('./routes/profile');
+const missionsRoutes = require('./routes/missions');
+const classesRoutes = require('./routes/classes');
+const libraryRoutes = require('./routes/library');
+const badgesRoutes = require('./routes/badges');
+const pagesRoutes = require('./routes/pages');
+const navRoutes = require('./routes/nav');
+const agentRoutes = require('./routes/agent');
+const botLinkRoutes = require('./routes/bot-link');
+const { loadNavItems } = require('./util/nav-loader');
+
+const createApp = () => {
+  const app = express();
+
+  app.use(express.json());
+  app.use(express.urlencoded({ extended: true }));
+  app.use(express.static(path.join(__dirname, 'public')));
+
+  app.engine('handlebars', exphbs.engine({
+    layoutsDir: path.join(__dirname, 'views/layouts'),
+    partialsDir: path.join(__dirname, 'views/partials'),
+    defaultLayout: 'main',
+    helpers: {
+      ...helpers,
+      times,
+      range,
+      date_tz,
+      calendar_link,
+      encodeURIComponentH,
+      getTotalV1MissionsNeeded,
+      getTotalV2MissionsNeeded,
+      setVariable,
+      dump,
+      videoEmbed,
+      isSupportedVideoUrl,
+      substring,
+      concat,
+      effectiveRulesVersion,
+      wordCount,
+      perksForAbility,
+      nextPerkPosition,
+      json,
+      markdown: renderMarkdown
+    }
+  }));
+  app.set('view engine', 'handlebars');
+  app.set('views', path.join(__dirname, 'views'));
+
+  app.use((req, res, next) => {
+    res.locals.supabaseUrl = process.env.SUPABASE_URL;
+    res.locals.supabaseKey = process.env.SUPABASE_PUBLISHABLE_KEY;
+    next();
+  });
+
+  app.use(loadNavItems);
+  app.use('/', homeRoutes);
+  app.use('/auth', authRoutes);
+  app.use('/profile', profileRoutes);
+  app.use('/characters', charactersRoutes);
+  app.use('/lfg', lfgRoutes);
+  app.use('/missions', missionsRoutes);
+  app.use('/classes', classesRoutes);
+  app.use('/library', libraryRoutes);
+  app.use('/badges', badgesRoutes);
+  app.use('/pages', pagesRoutes);
+  app.use('/nav', navRoutes);
+  app.use('/link/bot', botLinkRoutes);
+  app.use('/api/agent', agentRoutes);
+
+  app.use((err, req, res, next) => {
+    console.error('Unhandled error:', err);
+    if (res.headersSent) return next(err);
+    return sendError(req, res, err);
+  });
+
+  return app;
+};
+
+module.exports = { createApp };
